[BZ 7517474] Initial support for tracking compliance
This commit adds a complianceStatus field to DriftDefinition. There are
corresponding dbsetup and dbupgrade changes. DriftManagerBean has been
updated to set the complianceStatus field if the definition is pinned
when the agent reports a change set.

+++ modules/core/plugin-container/src/main/java/org/rhq/core/pc/drift/DriftDetector.java
@@ -268,7 +268,7 @@ public class DriftDetector implements Runnable {
             // the current snapshot to match the pinned snapshot. Note though
             // that we increment the snapshot version in order to let the server
             // know about the state change.
-            if (schedule.getDriftDefinition().isPinned()) {
+            if (schedule.getDriftDefinition().isPinned() && newVersion > 1) {
                 currentSnapshot.delete();
                 File newSnapshot = updateCurrentSnapshot(schedule, snapshotEntries, newVersion);
